refactor: separando código em modulos
