+++ star-wars/index.html
@@ -18,7 +18,7 @@
 
     </div>
 
-    <script src="script.js"></script>
+    <script type="module" src="./scripts/index.js"></script>
 </body>
 
 </html>
